Move uniqueFeed method and LoaderStub class into a shared scope avoiding duplication
