+++ EssentialApp/EssentialAppTests/Helpers/SharedTestHelpers.swift
@@ -6,7 +6,7 @@
 //
 
 import Foundation
-
+import EssentialFeed
 
 func anyNSError() -> NSError {
     return NSError(domain: "any error", code: 0)
@@ -18,4 +18,11 @@ func anyURL() -> URL {
 
 func anyData() -> Data {
     return Data("any data".utf8)
+}
+
+func uniqueFeed() -> [FeedImage] {
+    return [FeedImage(id: UUID(),
+                      description: "any",
+                      location: "any",
+                      url: anyURL())]
 }
